feat: add SAT DOPPLER satellite tracking module
Real-time Doppler frequency correction for amateur radio satellites.
Standalone module (independent of spectrum analyzer).

Features:
- Automatic Doppler shift calculation using sigmoidal curve model
- Separate RX/TX correction (inverted for uplink compensation)
- Manual tuning in 1 kHz steps (keys 1/7 for RX, 3/9 for TX)
- Channel-based satellite database (channels named with $ prefix)
- Supports FM, USB (SSB), and AM modulation from channel config
- TX with CTCSS and power level from channel settings
- Monitor/squelch override with SIDE1
- Configurable pass duration before each tracking session
- No hardware modification or expanded EEPROM required

Activation: F+* from main screen

Technical:
- RTC driver using DP32G030 internal oscillator (lazy init, only when Doppler active)
- Minimal ISR (flag only), all calculation in main loop
- Doppler max shift = freq / 39972 (v_orbital/c approximation)
- T constant fixed at 1430 (500km LEO altitude)
- Retune only when shift exceeds 1 kHz to avoid audio glitches
- BK4819 squelch interrupt processing for proper RX behavior

Other changes:
- Welcome screen: shows "PU2DIY SAT DOPPLER", 4 second display time
- boot_counter_10ms changed to uint16_t for longer boot screen
- start.S: HandlerRTC declared as weak symbol for C override
- Makefile: added ENABLE_DOPPLER flag, -Wno-unterminated-string-initialization

Disabled to fit in flash: Messenger, F4HWN PMR/Narrower/Sleep

+++ src/radio/misc.c
@@ -305,7 +305,7 @@ volatile bool     gFlagTailNoteEliminationComplete;
     volatile bool gScheduleFM;
 #endif
 
-volatile uint8_t  boot_counter_10ms;
+volatile uint16_t boot_counter_10ms;
 
 uint8_t           gIsLocked = 0xFF;
